+++ src/DomainModule/JSON/JSDM_SimpleData.cpp
@@ -5,7 +5,7 @@ namespace Nebulite::DomainModule::JSON {
 
 //------------------------------------------
 // Update
-Nebulite::Constants::Error SimpleData::update() {
+Nebulite::Constants::Error SimpleData::update(){
     std::lock_guard<std::recursive_mutex> mtx = domain->lock(); // Lock the domain for thread-safe access
     // Add Domain-specific updates here!
     // General rule:
@@ -19,16 +19,16 @@ Nebulite::Constants::Error SimpleData::update() {
 //------------------------------------------
 // General set/get/remove functions
 
-Nebulite::Constants::Error SimpleData::set(int argc,  char* argv[]) {
+Nebulite::Constants::Error SimpleData::set(int argc,  char* argv[]){
     std::lock_guard<std::recursive_mutex> mtx = domain->lock(); // Lock the domain for thread-safe access
-    if(argc < 3) {
+    if(argc < 3){
         Nebulite::Utility::Capture::cerr() << "Error: Too few arguments for set command." << Nebulite::Utility::Capture::endl;
         return Nebulite::Constants::ErrorTable::FUNCTIONAL::TOO_FEW_ARGS();
     }
     
     std::string key = argv[1];
     std::string value = argv[2];
-    for (int i = 3; i < argc; ++i) {
+    for (int i = 3; i < argc; ++i){
         value += " " + std::string(argv[i]);
     }
     domain->set(key.c_str(), value);
@@ -44,7 +44,7 @@ Note: All values are stored as strings.
 
 Nebulite::Constants::Error SimpleData::move(int argc,  char* argv[]){
     std::lock_guard<std::recursive_mutex> mtx = domain->lock(); // Lock the domain for thread-safe access
-    if (argc != 3) {
+    if (argc != 3){
         Nebulite::Utility::Capture::cerr() << "Error: Too few arguments for move command." << Nebulite::Utility::Capture::endl;
         return Nebulite::Constants::ErrorTable::FUNCTIONAL::TOO_FEW_ARGS();
     }
@@ -52,22 +52,22 @@ Nebulite::Constants::Error SimpleData::move(int argc,  char* argv[]){
     std::string sourceKey = argv[1];
     std::string targetKey = argv[2];
 
-    if(domain->memberCheck(sourceKey) == Nebulite::Utility::JSON::KeyType::null) {
+    if(domain->memberCheck(sourceKey) == Nebulite::Utility::JSON::KeyType::null){
         Nebulite::Utility::Capture::cerr() << "Error: Source key '" << sourceKey << "' does not exist." << Nebulite::Utility::Capture::endl;
         return Nebulite::Constants::ErrorTable::FUNCTIONAL::UNKNOWN_ARG();
     }
-    if(domain->memberCheck(sourceKey) == Nebulite::Utility::JSON::KeyType::document) {
+    if(domain->memberCheck(sourceKey) == Nebulite::Utility::JSON::KeyType::document){
         Nebulite::Utility::JSON subdoc = domain->get_subdoc(sourceKey);
         domain->remove_key(targetKey.c_str());
         domain->set_subdoc(targetKey.c_str(), &subdoc);
         domain->remove_key(sourceKey.c_str());
     }
-    else if (domain->memberCheck(sourceKey) == Nebulite::Utility::JSON::KeyType::array) {
+    else if (domain->memberCheck(sourceKey) == Nebulite::Utility::JSON::KeyType::array){
         // Careful handling required:
         domain->remove_key(targetKey.c_str());
 
-        uint16_t size = domain->memberSize(sourceKey);
-        for (uint16_t i = 0; i < size; ++i) {
+        size_t size = domain->memberSize(sourceKey);
+        for (size_t i = 0; i < size; ++i){
             std::string itemKey = sourceKey + "[" + std::to_string(i) + "]";
             std::string itemValue = domain->get<std::string>(itemKey.c_str());
             std::string targetItemKey = targetKey + "[" + std::to_string(i) + "]";
@@ -91,7 +91,7 @@ Usage: move <source_key> <destination_key>
 
 Nebulite::Constants::Error SimpleData::copy(int argc,  char* argv[]){
     std::lock_guard<std::recursive_mutex> mtx = domain->lock(); // Lock the domain for thread-safe access
-    if (argc != 3) {
+    if (argc != 3){
         Nebulite::Utility::Capture::cerr() << "Error: Too few arguments for copy command." << Nebulite::Utility::Capture::endl;
         return Nebulite::Constants::ErrorTable::FUNCTIONAL::TOO_FEW_ARGS();
     }
@@ -99,21 +99,21 @@ Nebulite::Constants::Error SimpleData::copy(int argc,  char* argv[]){
     std::string sourceKey = argv[1];
     std::string targetKey = argv[2];
 
-    if(domain->memberCheck(sourceKey) == Nebulite::Utility::JSON::KeyType::null) {
+    if(domain->memberCheck(sourceKey) == Nebulite::Utility::JSON::KeyType::null){
         Nebulite::Utility::Capture::cerr() << "Error: Source key '" << sourceKey << "' does not exist." << Nebulite::Utility::Capture::endl;
         return Nebulite::Constants::ErrorTable::FUNCTIONAL::UNKNOWN_ARG();
     }
-    if(domain->memberCheck(sourceKey) == Nebulite::Utility::JSON::KeyType::document) {
+    if(domain->memberCheck(sourceKey) == Nebulite::Utility::JSON::KeyType::document){
         Nebulite::Utility::JSON subdoc = domain->get_subdoc(sourceKey);
         domain->remove_key(targetKey.c_str());
         domain->set_subdoc(targetKey.c_str(), &subdoc);
     }
-    else if (domain->memberCheck(sourceKey) == Nebulite::Utility::JSON::KeyType::array) {
+    else if (domain->memberCheck(sourceKey) == Nebulite::Utility::JSON::KeyType::array){
         // Careful handling required:
         domain->remove_key(targetKey.c_str());
 
         uint16_t size = domain->memberSize(sourceKey);
-        for (uint16_t i = 0; i < size; ++i) {
+        for (uint16_t i = 0; i < size; ++i){
             std::string itemKey = sourceKey + "[" + std::to_string(i) + "]";
             std::string itemValue = domain->get<std::string>(itemKey.c_str());
             std::string targetItemKey = targetKey + "[" + std::to_string(i) + "]";
@@ -136,7 +136,7 @@ Usage: copy <source_key> <destination_key>
 
 Nebulite::Constants::Error SimpleData::keyDelete(int argc,  char* argv[]){
     std::lock_guard<std::recursive_mutex> mtx = domain->lock(); // Lock the domain for thread-safe access
-    if (argc != 2) {
+    if (argc != 2){
         Nebulite::Utility::Capture::cerr() << "Error: Too few arguments for delete command." << Nebulite::Utility::Capture::endl;
         return Nebulite::Constants::ErrorTable::FUNCTIONAL::TOO_FEW_ARGS();
     }
@@ -152,13 +152,13 @@ Usage: keyDelete <key>
 
 //------------------------------------------
 // Array manipulation functions
-Nebulite::Constants::Error SimpleData::ensureArray(int argc,  char* argv[]) {
+Nebulite::Constants::Error SimpleData::ensureArray(int argc,  char* argv[]){
     std::lock_guard<std::recursive_mutex> mtx = domain->lock(); // Lock the domain for thread-safe access
-    if (argc < 2) {
+    if (argc < 2){
         Nebulite::Utility::Capture::cerr() << "Error: Too few arguments for ensureArray command." << Nebulite::Utility::Capture::endl;
         return Nebulite::Constants::ErrorTable::FUNCTIONAL::TOO_FEW_ARGS();
     }
-    if (argc > 2) {
+    if (argc > 2){
         Nebulite::Utility::Capture::cerr() << "Error: Too many arguments for ensureArray command." << Nebulite::Utility::Capture::endl;
         return Nebulite::Constants::ErrorTable::FUNCTIONAL::TOO_MANY_ARGS();
     }
@@ -167,12 +167,12 @@ Nebulite::Constants::Error SimpleData::ensureArray(int argc,  char* argv[]) {
 
     Nebulite::Utility::JSON::KeyType keyType = domain->memberCheck(key);
 
-    if (keyType == Nebulite::Utility::JSON::KeyType::array) {
+    if (keyType == Nebulite::Utility::JSON::KeyType::array){
         // Already an array, nothing to do
         return Nebulite::Constants::ErrorTable::NONE();
     }
 
-    if(keyType == Nebulite::Utility::JSON::KeyType::value) {
+    if(keyType == Nebulite::Utility::JSON::KeyType::value){
         // pop out value
         std::string existingValue = domain->get<std::string>(key.c_str());
         domain->remove_key(key.c_str());
@@ -196,7 +196,7 @@ Usage: ensure-array <key>
 
 Nebulite::Constants::Error SimpleData::push_back(int argc,  char* argv[]){
     std::lock_guard<std::recursive_mutex> mtx = domain->lock(); // Lock the domain for thread-safe access
-    if (argc > 3) {
+    if (argc > 3){
         Nebulite::Utility::Capture::cerr() << "Error: Too many arguments for push_front command." << Nebulite::Utility::Capture::endl;
         return Nebulite::Constants::ErrorTable::FUNCTIONAL::TOO_MANY_ARGS();
     }
@@ -211,12 +211,12 @@ Nebulite::Constants::Error SimpleData::push_back(int argc,  char* argv[]){
         value = argv[2];
     }
 
-    if (domain->memberCheck(key) != Nebulite::Utility::JSON::KeyType::array) {
+    if (domain->memberCheck(key) != Nebulite::Utility::JSON::KeyType::array){
         std::string command = __FUNCTION__;
         command += " " + SimpleData::ensureArray_name;
         command += " " + key;
         Nebulite::Constants::Error result = domain->parseStr(command);
-        if (result != Nebulite::Constants::ErrorTable::NONE()) {
+        if (result != Nebulite::Constants::ErrorTable::NONE()){
             Nebulite::Utility::Capture::cerr() << "Error: Failed to ensure array for key '" << key << "'." << Nebulite::Utility::Capture::endl;
             return result;
         }
@@ -235,29 +235,29 @@ Usage: push-back <key> <value>
 
 Nebulite::Constants::Error SimpleData::pop_back(int argc,  char* argv[]){
     std::lock_guard<std::recursive_mutex> mtx = domain->lock(); // Lock the domain for thread-safe access
-    if (argc < 2) {
+    if (argc < 2){
         Nebulite::Utility::Capture::cerr() << "Error: Too few arguments for push_back command." << Nebulite::Utility::Capture::endl;
         return Nebulite::Constants::ErrorTable::FUNCTIONAL::TOO_FEW_ARGS();
     }
-    if (argc > 2) {
+    if (argc > 2){
         Nebulite::Utility::Capture::cerr() << "Error: Too many arguments for push_back command." << Nebulite::Utility::Capture::endl;
         return Nebulite::Constants::ErrorTable::FUNCTIONAL::TOO_MANY_ARGS();
     }
     std::string key = argv[1];
 
-    if (domain->memberCheck(key) != Nebulite::Utility::JSON::KeyType::array) {
+    if (domain->memberCheck(key) != Nebulite::Utility::JSON::KeyType::array){
         std::string command = __FUNCTION__;
         command += " " + SimpleData::ensureArray_name;
         command += " " + key;
         Nebulite::Constants::Error result = domain->parseStr(command);
-        if (result != Nebulite::Constants::ErrorTable::NONE()) {
+        if (result != Nebulite::Constants::ErrorTable::NONE()){
             Nebulite::Utility::Capture::cerr() << "Error: Failed to ensure array for key '" << key << "'." << Nebulite::Utility::Capture::endl;
             return result;
         }
     }
 
     size_t size = domain->memberSize(key);
-    if (size == 0) {
+    if (size == 0){
         // nothing to pop out, not seen as error
         return Nebulite::Constants::ErrorTable::NONE();
     }
@@ -274,7 +274,7 @@ Usage: pop-back <key>
 
 Nebulite::Constants::Error SimpleData::push_front(int argc,  char* argv[]){
     std::lock_guard<std::recursive_mutex> mtx = domain->lock(); // Lock the domain for thread-safe access
-    if (argc > 3) {
+    if (argc > 3){
         Nebulite::Utility::Capture::cerr() << "Error: Too many arguments for push_front command." << Nebulite::Utility::Capture::endl;
         return Nebulite::Constants::ErrorTable::FUNCTIONAL::TOO_MANY_ARGS();
     }
@@ -290,12 +290,12 @@ Nebulite::Constants::Error SimpleData::push_front(int argc,  char* argv[]){
     }
 
     
-    if (domain->memberCheck(key) != Nebulite::Utility::JSON::KeyType::array) {
+    if (domain->memberCheck(key) != Nebulite::Utility::JSON::KeyType::array){
         std::string command = __FUNCTION__;
         command += " " + SimpleData::ensureArray_name;
         command += " " + key;
         Nebulite::Constants::Error result = domain->parseStr(command);
-        if (result != Nebulite::Constants::ErrorTable::NONE()) {
+        if (result != Nebulite::Constants::ErrorTable::NONE()){
             Nebulite::Utility::Capture::cerr() << "Error: Failed to ensure array for key '" << key << "'." << Nebulite::Utility::Capture::endl;
             return result;
         }
@@ -307,10 +307,10 @@ Nebulite::Constants::Error SimpleData::push_front(int argc,  char* argv[]){
     // Security check:
     // if any array item is a document, throw error
     // This feature is yet to be implemented!
-    for (size_t i = 0; i < size; ++i) {
+    for (size_t i = 0; i < size; ++i){
         std::string itemKey = key + "[" + std::to_string(i) + "]";
         Nebulite::Utility::JSON::KeyType itemType = domain->memberCheck(itemKey);
-        if (itemType == Nebulite::Utility::JSON::KeyType::document) {
+        if (itemType == Nebulite::Utility::JSON::KeyType::document){
             Nebulite::Utility::Capture::cerr() << "Error: Cannot push_front into an array containing documents." << Nebulite::Utility::Capture::endl;
             return Nebulite::Constants::ErrorTable::FUNCTIONAL::CRITICAL_FUNCTION_NOT_IMPLEMENTED();
         }
@@ -318,7 +318,7 @@ Nebulite::Constants::Error SimpleData::push_front(int argc,  char* argv[]){
 
     //------------------------------------------
     // Move all existing items one step forward
-    for (size_t i = size; i > 0; --i) {
+    for (size_t i = size; i > 0; --i){
         std::string itemKey = key + "[" + std::to_string(i - 1) + "]";
         std::string itemValue = domain->get<std::string>(itemKey.c_str());
         std::string newItemKey = key + "[" + std::to_string(i) + "]";
@@ -336,22 +336,22 @@ Usage: push-front <key> <value>
 
 Nebulite::Constants::Error SimpleData::pop_front(int argc,  char* argv[]){
     std::lock_guard<std::recursive_mutex> mtx = domain->lock(); // Lock the domain for thread-safe access
-    if (argc < 2) {
+    if (argc < 2){
         Nebulite::Utility::Capture::cerr() << "Error: Too few arguments for pop_front command." << Nebulite::Utility::Capture::endl;
         return Nebulite::Constants::ErrorTable::FUNCTIONAL::TOO_FEW_ARGS();
     }
-    if (argc > 2) {
+    if (argc > 2){
         Nebulite::Utility::Capture::cerr() << "Error: Too many arguments for pop_front command." << Nebulite::Utility::Capture::endl;
         return Nebulite::Constants::ErrorTable::FUNCTIONAL::TOO_MANY_ARGS();
     }
     std::string key = argv[1];
 
-    if (domain->memberCheck(key) != Nebulite::Utility::JSON::KeyType::array) {
+    if (domain->memberCheck(key) != Nebulite::Utility::JSON::KeyType::array){
         std::string command = __FUNCTION__;
         command += " " + SimpleData::ensureArray_name;
         command += " " + key;
         Nebulite::Constants::Error result = domain->parseStr(command);
-        if (result != Nebulite::Constants::ErrorTable::NONE()) {
+        if (result != Nebulite::Constants::ErrorTable::NONE()){
             Nebulite::Utility::Capture::cerr() << "Error: Failed to ensure array for key '" << key << "'." << Nebulite::Utility::Capture::endl;
             return result;
         }
@@ -363,10 +363,10 @@ Nebulite::Constants::Error SimpleData::pop_front(int argc,  char* argv[]){
     // Security check:
     // if any array item is a document, throw error
     // This feature is yet to be implemented!
-    for (size_t i = 0; i < size; ++i) {
+    for (size_t i = 0; i < size; ++i){
         std::string itemKey = key + "[" + std::to_string(i) + "]";
         Nebulite::Utility::JSON::KeyType itemType = domain->memberCheck(itemKey);
-        if (itemType == Nebulite::Utility::JSON::KeyType::document) {
+        if (itemType == Nebulite::Utility::JSON::KeyType::document){
             Nebulite::Utility::Capture::cerr() << "Error: Cannot push_front into an array containing documents." << Nebulite::Utility::Capture::endl;
             return Nebulite::Constants::ErrorTable::FUNCTIONAL::CRITICAL_FUNCTION_NOT_IMPLEMENTED();
         }
@@ -374,7 +374,7 @@ Nebulite::Constants::Error SimpleData::pop_front(int argc,  char* argv[]){
 
     //------------------------------------------
     // Move all existing items one step back
-    for (size_t i = 1; i < size; i++) {
+    for (size_t i = 1; i < size; i++){
         std::string itemKey = key + "[" + std::to_string(i) + "]";
         std::string itemValue = domain->get<std::string>(itemKey.c_str());
         std::string newItemKey = key + "[" + std::to_string(i - 1) + "]";
